requirements.txt
Django==1.11.7
django-ckeditor==5.3.1
django-form-utils==1.0.3
django-js-asset==0.1.1
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
pytz==2017.3